requirements.txt
Cython>=0.19.2
numpy>=1.7.1
scipy>=0.13.2
scikit-image>=0.9.3
matplotlib>=1.3.1
ipython>=3.0.0
pyyaml>=3.10
Pillow>=2.3.0
easydict>=1.6
ijson>=2.3
tqdm>=4.17.1